% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustBroaden.R
\name{AdjustBroaden}
\alias{AdjustBroaden}
\title{Processes spectrum to correct FourXP format}
\usage{
AdjustBroaden(spec)
}
\arguments{
\item{spec}{1D vector of spectrum fluxes}
}
\description{
Broaden spectrum by 1 pixel using a maximum filter kernel of width 3.
}
\examples{
None applicable
}
\author{
I. Baldry, L. Davies, L Drygala
}
