% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosineFilter.R
\name{CosineFilter}
\alias{CosineFilter}
\title{CosineFilter}
\usage{
CosineFilter(num, reverse = FALSE)
}
\arguments{
\item{num}{number of cosine filter}

\item{reverse}{TRUE/FALSE reverse filter}
}
\description{
Rising or falling cosine filter, i.e., a cosine bell taper.
}
\examples{
None applicable
}
\author{
I. Baldry, L. Davies, L Drygala
}
