% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriteriaBetween.R
\name{CriteriaBetween}
\alias{CriteriaBetween}
\title{CriteriaBetween}
\usage{
CriteriaBetween(data, data_range)
}
\arguments{
\item{data}{1D vector of data}

\item{data_range}{data range to return 1 within.}
}
\description{
Return 1 where data is within range
}
\examples{
None applicable
}
\author{
I. Baldry, L. Davies, L Drygala
}
