% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriteriaPeak.R
\name{CriteriaPeak}
\alias{CriteriaPeak}
\title{CriteriaPeak}
\usage{
CriteriaPeak(values)
}
\arguments{
\item{values}{1D data values}
}
\description{
Determine where local peak values are in data.
}
\examples{
None applicable as internal function.... 
}
\author{
I. baldry, L. Davies <luke.j.davies@uwa.edu>
}
