% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossCorr.R
\name{CrossCorr}
\alias{CrossCorr}
\title{CrossCorr}
\usage{
CrossCorr(template, spec, plan = plan, corrSize = 14000)
}
\arguments{
\item{template}{1D template spectrum}

\item{spec}{1D input spectrum}

\item{plan}{FFT plan}

\item{corrSize}{size of output correlation vector}
}
\description{
Cross correlate one spectrum with a template
}
\examples{
None applicable as internal function.... 
}
\author{
I. baldry, L. Davies <luke.j.davies@uwa.edu>
}
