% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlations_old.R, R/DoCrossCorr.R
\name{DoCrossCorr}
\alias{DoCrossCorr}
\title{Compute cross corelations in FourXP_Z}
\usage{
DoCrossCorr(spec, gap, tempData, helioVel, plan, z_prior, highZ = FALSE)

DoCrossCorr(spec, gap, tempData, helioVel, plan, z_prior, highZ = FALSE)
}
\arguments{
\item{spec}{input spectrum containing spec$lambda and spec$flux}

\item{gap}{gap in shifts probed in cross correlation}

\item{plan}{fast-forier tranform plan}

\item{z_prior}{redshift priors, c(lo,hi)}

\item{tempDatar}{structure of template data}

\item{heloval}{helocentric correction value}

\item{spec}{input spectrum containing spec$lambda and spec$flux}

\item{gap}{gap in shifts probed in cross correlation}

\item{tempDatar}{structure of template data}

\item{heloval}{helocentric correction value}

\item{plan}{fast-forier tranform plan}

\item{z_prior}{redshift priors, c(lo,hi)}
}
\description{
Function to compute cross-correlation strngth between a given spectrum and
a set of templates. Function is internal to FourXP_Z.

Function to compute cross-correlation strngth between a given spectrum and
a set of templates. Function is internal to FourXP_Z.
}
\examples{
None applicable as internal function.... 
None applicable as internal function.... 
}
\author{
I. baldry, L. Davies <luke.j.davies@uwa.edu>

L. Davies <luke.j.davies@uwa.edu>
}
