% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitandFilter.R
\name{FitandFilter}
\alias{FitandFilter}
\title{FitandFilter}
\usage{
FitandFilter(spec, stLambda, endLambda, os1, os2, verbose = TRUE)
}
\arguments{
\item{spec}{list with elements spec$flux and spec$wave}

\item{stLambda}{starting wavelength for fitting}

\item{endLambda}{ending wavelength for fitting}

\item{os1}{tmp}

\item{os2}{tmp}

\item{verbose}{TRUE/FALSE tell me what's going on}
}
\description{
Set up region to use for fitting between start Lambda and end Lambda.
}
\examples{
None applicable
}
\author{
I. Baldry, L. Davies, L Drygala
}
