% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourXP_Sim.R
\name{FourXP_Sim}
\alias{FourXP_Sim}
\title{Simulated and mock observe spectra with 4MOST}
\usage{
FourXP_Sim(id = "Test", zIn = 0.234, mag = 19.8, band = "VST_r",
  col = 0.55, mass = 10.2, sfr = 7, agn = "F",
  specDir = "/Users/luke/work/IWG8/FourXPmodels/", expMin = 60,
  nSub = 3, SKYBRIGHT_TYPE = "ZENITH", AIRMASS = 1.4, IQ = 1.1,
  SKYBRIGHT = 21.77, TILT = 6, MISALIGNMENT = 0.1,
  systemModDir = "/Applications/4FS-ETC_app/4FS_ETC_system_model_v0.2/",
  plot = T, verbose = TRUE)
}
\arguments{
\item{id}{ID to assing to this simulation}

\item{zIn}{input redshift to simulate}

\item{mag}{magnitude to scale final spectrum to}

\item{band}{photometric band in which scaling is to be done. See ?getfilt for options}

\item{col}{g-i colour to match to template spectrum}

\item{mass}{Log10(stellar mass) to match to templale spectrum}

\item{sfr}{SFR to match to templale spectrum}

\item{agn}{Use AGN template? 'F'=no AGN template, 'B'= broad-line template, 'N'= narrow-line template}

\item{specDir}{location of model database, this is required but is somewhat large (~2.5Gb). As such,it can be found here: https://www.dropbox.com/s/hwnbabirn59796x/FourXPmodels.zip?dl=0}

\item{expMin}{Total exposure time to simulate observation}

\item{nSub}{number of sub-exposure in expMin}

\item{SKYBRIGHT_TYPE}{where to measuresky brightness ZENITH or LOCAL?}

\item{AIRMASS}{simulated airmass of observation}

\item{IQ}{image quality value to simulate (V-band,FWHM,arcsec)}

\item{SKYBRIGHT}{sky brightness in mag/arcsec^2}

\item{TILT}{fibre tilt in mm}

\item{MISALIGNMENT}{fibre misallignment in arcsec}

\item{systemModDir}{IMPORTANT: directory location of 4FS system files. Will be something like '4FS-ETC_app/4FS_ETC_system_model_v0.2/'}

\item{plot}{TRUE/FALSE plot the output spectrum with plotSpec()}

\item{verbose}{TRUE/FALSE let me know what's going on}
}
\value{
a list contining lots and lots of things from the simulated spectrum
}
\description{
High level function that combines most of the functionality of FourXP. 
Allows the user to generate a simulated spectrum and 'observe' with the 4MOST ETC. 
Various statisitics about the observation are returned.
}
\examples{
out<-FourXP_Sim(id='Test', zIn=0.234, mag=19.8, band='VST_r', col=0.55, mass=10.2,sfr=7, agn='F', expMin=60, nSub=3, verbose=TRUE)
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
