% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourXP_Zpat.R
\name{FourXP_Zpat}
\alias{FourXP_Zpat}
\title{Test version of redshfiting code for emission line pattern matching}
\usage{
FourXP_Zpat(specRaw, z_prior = c(-1, 5), plotCorr = T,
  verbose = TRUE, filter = TRUE)
}
\arguments{
\item{specRaw}{Input spectrum, list in format of all of the things in FourXP}

\item{z_prior}{redshift prior}

\item{plotCorr}{TRUE/FALSE plot the output of the match}

\item{verbose}{TRUE/FALSE let me know what's going on}

\item{filter}{TRUE/FALSE filter the spectrum for noise line peaks}
}
\value{
a list contining lots and lots of things from the simulated spectrum
}
\description{
Identifies emission lines in target spectrum and 
pattern matches to a catlogue of known emission line.
}
\examples{
spec<-makeSpec(id='D134', z=zIn, mag=21, band='VST_r', col=0.6, mass=11,sfr=10, agn='F')
zFit<-FourXP_Zpat(spec, z_prior=c(zIn-0.1,zIn+0.1), plotCorr=T)
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
