% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.spec.R
\name{get.spec}
\alias{get.spec}
\title{Read in spectra from FITS files}
\usage{
get.spec(file, row = 1, logw = F, sc = 1, xunit = "ang",
  yunit = "ang", z = NA, RA = NA, DEC = NA, errorRow = NA,
  SDSS_spec = F)
}
\arguments{
\item{file}{file name of FITS file to read}

\item{row}{row number to read in}

\item{logw}{TRUE/FALSE is file in log wavelength}

\item{sc}{value to scale inout fluxes}

\item{xunit}{xunit value of spectrum (ang, hz, micron,nm,m)}

\item{yunit}{xunit value of spectrum (ang=ergs/sec/cm^2/ang, hz=ergs/sec/cm^2/hz, Jy)}

\item{z}{input redshfit for spectrum (tries to read this from header)}

\item{RA}{input RA for spectrum (tries to read this from header)}

\item{DEC}{input DEC for spectrum (tries to read this from header)}

\item{errorRow}{row in spectrum of error (if available)}

\item{SDSS_spec}{TRUE/FALSE is this an SDSS spectrum (uses specific format)}
}
\value{
List containing wave,flux,xunit,yunit, z, RA, DEC, error
}
\description{
Reads in spectra from FITS files in a sensible-ish way and make list useable by
 other functions in FourXP.
}
\examples{
name<-paste(specDir,'Spectra/AGN_v1.0_z050_mr165_type1.fits',sep='')
spec<-get.spec(name, xunit='ang', yunit='ang', z=0.5)
plotSpec(spec, Air=F)
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
