% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRange.R
\name{GetRange}
\alias{GetRange}
\title{GetRange}
\usage{
GetRange(tNum, z_prior)
}
\arguments{
\item{tNum}{input spectrum containing spec$lambda and spec$flux}

\item{z_prior}{user-define photo-z prior}
}
\description{
Set allowable redshfit ranges of AutoZ Templates
}
\examples{
None applicable as internal function.... 
}
\author{
I. baldry, L. Davies <luke.j.davies@uwa.edu>
}
