% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install4XP.R
\name{install4XP}
\alias{install4XP}
\title{Install/Download libraries for 4XP}
\usage{
install4XP(downloadModels = TRUE, ModelDir = ".")
}
\arguments{
\item{downloadModels}{TRUE/FALSE download models, if FALSE calibrators will just be unpacked}

\item{ModelDir}{where to unpack FourXP models}
}
\value{
SpecDir (this will need to be used as a variable in FourXP_Sim() and makeSpec())
}
\description{
Downlaod and unpack all things required for
}
\examples{
install4XP()
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
