% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpec.R
\name{makeSpec}
\alias{makeSpec}
\title{Make a simulated spectrum using a set of SDSS+AGN templates at Air wavelengths}
\usage{
makeSpec(id = "tmp", z = 0.456, mag = 19.8, band = "VST_r",
  col = 0.4, mass = 10.2, sfr = 10, agn = "F",
  specDir = "/Users/luke/work/IWG8/FourXPmodels/")
}
\arguments{
\item{id}{id to assign to spectrum}

\item{z}{redshfit for spectrum you wish to simulate}

\item{mag}{magnitude to scale final spectrum to}

\item{band}{photometric band in which scaling is to be done. See ?getfilt for options}

\item{col}{g-i colour to match to template spectrum}

\item{mass}{Log10(stellar mass) to match to templale spectrum}

\item{sfr}{SFR to match to templale spectrum}

\item{agn}{Use AGN template? 'N'=no AGN template, 'B'= broad-line template, 'N'= narrow-line template}

\item{specDir}{location of model database - link to this will have been provided by running install4XP()}
}
\description{
Function takes inputs of redshift, magnitude, band for magnitude, g-i colour, stellar mass, 
sfr and AGN-type and returns a list with the paramters for a simulated spectrum. If AGN=='F': g-i colour, stellar mass 
and SFR are matched to a set of PPXF fits to SDSS spectra in the GAMA regions, the best-matched source spectrum is used.
This is then fit to the BC03 continuum models to extend in wavelength range. Note that the final flux sacling is not 
defined by col, mass and SFR, only the spectral type. If AGN='B'/'N', col, mass and SFR are ignored and either a 
broad-line ('B') or narrow ('N') line tempalte is used. This is still scaled to the correct magnitude and redshift.
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
