% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanReject.R
\name{MeanReject}
\alias{MeanReject}
\title{MeanReject}
\usage{
MeanReject(data, numReject)
}
\arguments{
\item{data}{1D data values}

\item{numReject}{number of value to reject}
}
\description{
Take the mean of a set of values after rejecting numReject lowest
}
\examples{
None applicable as internal function.... 
}
\author{
I. baldry, L. Davies <luke.j.davies@uwa.edu>
}
