% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianAdjust.R
\name{MedianAdjust}
\alias{MedianAdjust}
\title{MedianAdjust}
\usage{
MedianAdjust(x, width)
}
\arguments{
\item{x}{1D vector to be smoothed}

\item{width}{width of smoothing kernal}
}
\description{
# Perform median filter with edge adjustment on a vector spectrum.
}
\examples{
None applicable
}
\author{
I. Baldry, L. Davies, L Drygala
}
