% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedSmooth.R
\name{MedSmooth}
\alias{MedSmooth}
\title{MedSmooth}
\usage{
MedSmooth(spec, medWidth = 51, smoothWidth = 121)
}
\arguments{
\item{spec}{1D vector to be filtered and smoothed}

\item{medWidth}{width of filtering kernal}

\item{smoothWidth}{width of smoothing kernal}
}
\description{
# Median filter a spectrum then smooth.
}
\examples{
None applicable
}
\author{
I. Baldry, L. Davies, L Drygala
}
