% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormaliseMeandev.R
\name{NormaliseMeandev}
\alias{NormaliseMeandev}
\title{NormaliseMeandev}
\usage{
NormaliseMeandev(flux, clipvalue, use, percentvalue = 0.95)
}
\arguments{
\item{flux}{1D vector of spectrum fluxes}

\item{clipvalue}{Value to clip/replace flux}

\item{use}{1D vector of elements in flux to use}

\item{percentvalue}{is 'use' not defined, quantile of value overwehich to ignore}
}
\description{
Normalise a spectrum by dividing by mean absolute deviation. Iterate for clipping if clipvalue keyword is set.
}
\examples{
None applicable
}
\author{
I. Baldry, L. Davies, L Drygala
}
