% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observeSpec4FS.R
\name{observeSpec4FS}
\alias{observeSpec4FS}
\title{Observe a spectrum with 4FS ETC}
\usage{
observeSpec4FS(spec, expMin = 120, nSub = 4,
  SKYBRIGHT_TYPE = "ZENITH", AIRMASS = 1.4, IQ = 1.1,
  SKYBRIGHT = 21.77, TILT = 6, MISALIGNMENT = 0.1,
  systemModDir = "/Applications/4FS-ETC_app/4FS_ETC_system_model_v0.2/",
  keepFITS = TRUE)
}
\arguments{
\item{spec}{input list contianing spectral information. Minimum requirements are
wave=vector of wavelengths in angstrom and flux=vector of fluxes in erg/s/cm2/Angstrom.
This is ideally the output of makeSpec(), which contains all of the relavent paramters.}

\item{expMin}{Total exposure time to simulate observation}

\item{nSub}{number of sub-exposure in expMin}

\item{SKYBRIGHT_TYPE}{where to measuresky brightness ZENITH or LOCAL?}

\item{AIRMASS}{simulated airmass of observation}

\item{IQ}{image quality value to simulate (V-band,FWHM,arcsec)}

\item{SKYBRIGHT}{sky brightness in mag/arcsec^2}

\item{TILT}{fibre tilt in mm}

\item{MISALIGNMENT}{fibre misallignment in arcsec}

\item{systemModDir}{IMPORTANT: directory location of 4FS system files. Will be something like '4FS-ETC_app/4FS_ETC_system_model_v0.2/'}

\item{keepFITS}{keep FITS files, etc generated by 4FS ETC?}
}
\value{
List containing lost fo things!
}
\description{
Take an input spectrum and mock observe using the 4MOST ETC. You must have
the 4MOST ETC software installed and executable from the command line using command '4FS_ETC'. If you
do not have this, contact L. Davies <luke.j.davies@uwa.edu> for further instructions.
Take in output of makeSpec() and a given set of paramters such a sexposure time, number of subexposures
and instrumental/observing set up from 4MOST. Returns a list with various observational paramters and 
can also save the ETC outputs and FITS files directly.
}
\examples{
spec<-makeSpec(id='D134', z=zTest, mag=19.8, band='VST_r', col=0.55, mass=10.2,sfr=7,agn='F')
observeSpec4FSOut<-observeSpec4FS(spec, expMin=60, nSub=3, keepFITS=F)
magplot(observeSpec4FSOut$blueRawWave, observeSpec4FSOut$blueRawFlux, type='l', col='blue', xlim=c(3500,10000), ylim=c(0, max(c(observeSpec4FSOut$blueRawFlux,observeSpec4FSOut$greenRawFlux,observeSpec4FSOut$redRawFlux),na.rm=T)), xlab='Wavelength, Ang', ylab='Counts')
lines(observeSpec4FSOut$greenRawWave, observeSpec4FSOut$greenRawFlux, type='l', col='darkgreen')
lines(observeSpec4FSOut$redRawWave, observeSpec4FSOut$redRawFlux, type='l', col='red')
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
