% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorssCorr.R
\name{plotCrossCorr}
\alias{plotCrossCorr}
\title{Plot the cross-correlation output from FourXP_Z()}
\usage{
plotCrossCorr(spec, templateNum = 13)
}
\arguments{
\item{spec}{R struture output from FourXP_Z()}

\item{templateNum}{Template number to plot correlation for}
}
\description{
Takes the output from FourXP_Z() and allows the user to plot the cross-correlation for a given template
}
\examples{
load(paste(.libPaths(),'/FourXP/data/ExampleSpec.Rdata',sep=''))
plot(spec$wave, hanning.smooth(spec$flux, degree=9), type='l', xlab='Wavelength, ang', ylab='Counts')
tempFile=paste(.libPaths(),'/FourXP/data/calibrators/AutoZTemp/filtered-templates.fits',sep='')
FourXP_Z_out<-FourXP_Z(spec, tempFile=tempFile, doHelio=F)
plotCrossCorr(FourXP_Z_out, templateNum=FourXP_Z_out$results[4])
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
