% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpec.basic.R
\name{plotSpec.basic}
\alias{plotSpec.basic}
\title{Plots a basic spectrum}
\usage{
plotSpec.basic(spec = spec, cex.axis = 1.4, cex.lab = 1.4,
  degSmooth = 7, xlim = NA, ylim = NA)
}
\arguments{
\item{spec}{spec object to plot}

\item{degSmooth}{Hanning smoothing to apply (odd number)}

\item{xlim}{x range of plot (if not supplied will use sensible values)}

\item{ylim}{y range of plot (if not supplied will use sensible values)}

\item{z}{FITS file row number}
}
\description{
Plots a basic spectrum with spec$flux and spec$wave parameters
}
\examples{
load(paste(.libPaths(),'/FourXP/data/ExampleSpec.Rdata',sep=''))
plotSpec.basic(spec=spec)
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
