% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyReject.R
\name{PolyReject}
\alias{PolyReject}
\title{PolyReject}
\usage{
PolyReject(x, y, niterate = 15, degree = 4, highSigma = 3.5,
  lowSigma = 3.5, verbose = TRUE)
}
\arguments{
\item{x}{1D vector of x values}

\item{y}{1D vector of y values}

\item{niterate}{number of iterations for sigma rejection}

\item{degree}{degree of polynomimial fit}

\item{highSigma}{sigma cut value}
}
\description{
Polynomial fitting with sigma rejection, return lm() object
}
\examples{
None applicable
}
\author{
I. Baldry, L. Davies, L Drygala
}
