% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSDSS.R
\name{readSDSS}
\alias{readSDSS}
\title{Read SDSS spectra}
\usage{
readSDSS(file, id = "test")
}
\arguments{
\item{file}{SDSS spectrum filename to read}

\item{id}{id to assing to spectum}
}
\description{
Reads in an SDSS sepctrum and converts to a sensibele format for 4XP
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
