% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RebinTempData.R
\name{RebinTempData}
\alias{RebinTempData}
\title{RebinTempData}
\usage{
RebinTempData(logLambda, templateNumbers = c(2:14, 16:22, 40:47),
  file = "data/calibrators/AutoZTemp/filtered-templates.fits ",
  verbose = TRUE)
}
\arguments{
\item{logLambda}{loglambda wavelength scale}

\item{templateNumbers}{template numbers to use}

\item{file}{templates file to use}

\item{verbose}{TRUE/FALSE tell me what's going on}
}
\description{
Rebin template data to same loglambda scale as data
}
\examples{
None applicable
}
\author{
I. Baldry, L. Davies
}
