% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningMeanSmooth.R
\name{RunningMeanSmooth}
\alias{RunningMeanSmooth}
\title{RunningMeanSmooth}
\usage{
RunningMeanSmooth(x, width)
}
\arguments{
\item{x}{1D vector to be smoothed}

\item{width}{width of smoothing kernal}
}
\description{
Translation of IDL's SMOOTH function into R Uses IDL's '/EDGE_TRUNCATE' option
}
\examples{
None applicable
}
\author{
L Drygala
}
