% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch4MOST.R
\name{stitch4MOST}
\alias{stitch4MOST}
\title{Stitches together 4MOST arms}
\usage{
stitch4MOST(blue_file, green_file, red_file, blue_thru_file,
  green_thru_file, red_thru_file, filt_scale = "VST_r", mag_scale = NA,
  makePlot = T, outName = "stictch4MOSTout.pdf")
}
\arguments{
\item{blue_file}{input FITS file for blue arm (output from 4MOST ETC)}

\item{green_file}{input FITS file for green arm (output from 4MOST ETC)}

\item{red_file}{input FITS file for red arm (output from 4MOST ETC)}

\item{blue_thru_file}{blue thougput file for ETC}

\item{green_thru_file}{green thougput file for ETC}

\item{red_thru_file}{red thougput file for ETC}

\item{filt_scale}{if you wish to scale to a magnitude in a particular band, name the band here as a string. Options can be found with ?getfilt}

\item{mag_scale}{AB magnitude in filt_scale to scale to}

\item{makePlot}{TRUE/FALSE make a PDF figure with the output spectrum and diagnositics}

\item{outName}{if makePlot==TRUE, string of output plot file names}
}
\value{
A list containing wave, flux, error, scaled flux, scaled error, mag_scale, filt_scale for spliced spectrum
}
\description{
Function takes outputs formthe 4MOST ETC and filter response curves and stiches them togther. 
Can also scale to a desired magnitude in a given band. However, to do this you will need FourXP::getfilt installed
and have unpacked the calibrator files:

> LibPaths<-.libPaths()

> system(paste('tar -xvf ', LibPaths, '/FourXP/data/calibrators.tar --directory ', LibPaths, '/FourXP/data/',sep=''))
}
\examples{
blue_file<-"specout_template_TempSim_SF_Vhighz_LRS_blue.fits"
green_file<-"specout_template_TempSim_SF_Vhighz_LRS_green.fits"
red_file<-"specout_template_TempSim_SF_Vhighz_LRS_red.fits"
blue_thru_file<-"4FS-ETC_app/4FS_ETC_system_model_v0.2/LRS/lrs_blue_material_4fs_efficiency_total.fits"
green_thru_file<-"4FS-ETC_app/4FS_ETC_system_model_v0.2/LRS/lrs_green_material_4fs_efficiency_total.fits"
red_thru_file<-"4FS-ETC_app/4FS_ETC_system_model_v0.2/LRS/lrs_red_material_4fs_efficiency_total.fits"
filt_scale='VST_r'
mag_scale=19
makeplot=T
outName='stictch4MOSTout.pdf'
spec<-stitch4MOST(blue_file,green_file,red_file, blue_thru_file, green_thru_file, red_thru_file, filt_scale=filt_scale, mag_scale=mag_scale, makePlot=T, outName=outName)
}
\author{
L. Davies <luke.j.davies@uwa.edu>
}
